Move less layout-related parts of typst_layout::grid to typst_library.
This will benefit HTML tables support, because it will use `CellGrid`,
but should not depend on typst_layout.
