+++ crates/typst-layout/src/grid/layouter.rs
@@ -12,6 +12,7 @@ use typst_library::visualize::Geometry;
 use typst_syntax::Span;
 use typst_utils::{MaybeReverseIter, Numeric};
 
+use super::layout_cell;
 use super::{
     generate_line_segments, hline_stroke_at_column, vline_stroke_at_row, Cell, CellGrid,
     LinePosition, LineSegment, Repeatable, Rowspan, UnbreakableRowGroup,
@@ -843,7 +844,8 @@ impl<'a> GridLayouter<'a> {
 
                 let size = Size::new(available, height);
                 let pod = Region::new(size, Axes::splat(false));
-                let frame = cell.layout(engine, 0, self.styles, pod.into())?.into_frame();
+                let frame =
+                    layout_cell(cell, engine, 0, self.styles, pod.into())?.into_frame();
                 resolved.set_max(frame.width() - already_covered_width);
             }
 
@@ -1086,7 +1088,7 @@ impl<'a> GridLayouter<'a> {
             };
 
             let frames =
-                cell.layout(engine, disambiguator, self.styles, pod)?.into_frames();
+                layout_cell(cell, engine, disambiguator, self.styles, pod)?.into_frames();
 
             // Skip the first region if one cell in it is empty. Then,
             // remeasure.
@@ -1252,9 +1254,9 @@ impl<'a> GridLayouter<'a> {
                         // rows.
                         pod.full = self.regions.full;
                     }
-                    let frame = cell
-                        .layout(engine, disambiguator, self.styles, pod)?
-                        .into_frame();
+                    let frame =
+                        layout_cell(cell, engine, disambiguator, self.styles, pod)?
+                            .into_frame();
                     let mut pos = pos;
                     if self.is_rtl {
                         // In the grid, cell colspans expand to the right,
@@ -1310,7 +1312,7 @@ impl<'a> GridLayouter<'a> {
 
                     // Push the layouted frames into the individual output frames.
                     let fragment =
-                        cell.layout(engine, disambiguator, self.styles, pod)?;
+                        layout_cell(cell, engine, disambiguator, self.styles, pod)?;
                     for (output, frame) in outputs.iter_mut().zip(fragment) {
                         let mut pos = pos;
                         if self.is_rtl {
