+++ crates/typst-library/src/model/mod.rs
@@ -1,5 +1,8 @@
 //! Structuring elements that define the document model.
 
+// TODO!
+pub mod grid;
+
 mod bibliography;
 mod cite;
 mod document;
